[CE-2404]: Recommendation - fix cost computation, misc feedback (#15336)
* [CE-2404]: Use limit if request is not available

* [CE-2404]: Fix cpu min (250m to 25m)

* [CE-2404]: Increase ttl

* [CE-2404]: ResourceId.of() -> builder

* [CE-2404]: Fix estimatedSavings computation

* [CE-2404]: Move K8S_WORKLOAD_RECOMMENDATION after INSTANCE_BILLING

* [CE-2404]: Add test coverage for cost computation

* [CE-2404]: Increase batch size

* [CE-2404]: Address sonar warnings

+++ 71-rest/src/main/java/software/wings/graphql/datafetcher/ce/recommendation/dto/QLK8sWorkloadRecommendation.java
@@ -6,6 +6,7 @@ import lombok.Value;
 import software.wings.security.PermissionAttribute;
 import software.wings.security.annotations.Scope;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Value
@@ -18,6 +19,6 @@ public class QLK8sWorkloadRecommendation {
   String clusterId;
   String clusterName;
   @Singular List<QLContainerRecommendation> containerRecommendations;
-  Double estimatedSavings;
+  BigDecimal estimatedSavings;
   int numDays;
 }
